# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Qwizzy API Documentation",
        "description": "Documentation de l'API Qwizzy pour la gestion des quiz",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/subjects": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "summary": "Obtenir la liste des sujets",
                "description": "Retourne la liste paginée des sujets",
                "operationId": "getSubjectsList",
                "parameters": [
                    {
                        "name": "current_sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "current_sort_dir",
                        "in": "query",
                        "description": "Direction du tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des sujets récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subject": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Subject"
                ],
                "summary": "Créer un nouveau sujet",
                "description": "Crée un nouveau sujet",
                "operationId": "storeSubject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 200,
                                        "example": "Mathématiques"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sujet créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/subjects/{id}": {
            "put": {
                "tags": [
                    "Subject"
                ],
                "summary": "Mettre à jour un sujet",
                "description": "Met à jour un sujet existant",
                "operationId": "updateSubject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du sujet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 200,
                                        "example": "Histoire"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sujet modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "404": {
                        "description": "Sujet non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subject"
                ],
                "summary": "Supprimer un sujet",
                "description": "Supprime un sujet existant",
                "operationId": "deleteSubject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du sujet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sujet supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Sujet non trouvé"
                    }
                }
            }
        }
    },
...
}